Struct aob::Match

pub struct Match<'haystack> { /* private fields */ }

Represents a matching Needle found in the haystack.

Implementations

impl<'haystack> Match<'haystack>

pub fn start(&self) -> usize

The position of the first byte in the matching needle, relative to the haystack.

let needle = DynamicNeedle::from_ida("63 ? 74").unwrap();
let haystack = "a_cat_tries";
let matched = needle.find(haystack.as_bytes()).unwrap();
assert_eq!(matched.start(), 2);

pub fn end(&self) -> usize

The position of the last byte past the end of the matching needle, relative to the haystack.

let needle = DynamicNeedle::from_ida("63 ? 74").unwrap();
let haystack = "a_cat_tries";
let matched = needle.find(haystack.as_bytes()).unwrap();
assert_eq!(matched.end(), 5);

pub fn range(&self) -> Range<usize>

The range of the matching needle, relative to the haystack.

let needle = DynamicNeedle::from_ida("63 ? 74").unwrap();
let haystack = "a_cat_tries";
let matched = needle.find(haystack.as_bytes()).unwrap();
assert_eq!(matched.range(), 2..5);

pub fn as_bytes(&self) -> &'haystack [u8]

The actual matched bytes, from the haystack.

let needle = DynamicNeedle::from_ida("63 ? 74").unwrap();
let haystack = "a_cat_tries";
let matched = needle.find(haystack.as_bytes()).unwrap();
assert_eq!(matched.as_bytes(), &b"cat"[..]);